Initial import
I've been maintaining this since 2011 as "mailsend", but this is pretty
much a rewrite using newer Python functionality that was not present
back then (much of the email functionality has been replaced by Python's
"email" implementation). Renamed the project to something that is not
available on PyPi so we can build a proper library.